+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Changed
 
+* Changed the parent class of the `MessageEvent` to `Symfony\Contracts\EventDispatcher\Event` when using `symfony/event-dispatcher` 4.3+
 * Support the new `symfony/event-dispatcher` 4.3+ signature
 
 ### Removed